병합으로 인한 중복 제거와 과제 해설자료 비교 일부 수정

+++ src/main/java/calculator/DivideOperator.java
@@ -3,7 +3,7 @@ package calculator;
 public class DivideOperator<Double> implements Operator<Double> {
 
     @Override
-    public double operate(Double a, Double b) {
-        return (double)a/(double)b;
+    public double operate(double a, double b) {
+        return a/b;
     }
 }
