+++ src/main/java/calculator/MultiplyOperator.java
@@ -2,7 +2,7 @@ package calculator;
 
 public class MultiplyOperator<Double> implements Operator<Double>  {
    @Override
-    public double operate(Double a, Double b) {
-        return (double)a * (double)b;
+    public double operate(double a, double b) {
+        return a*b;
     }
 }
